requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-rest-framework==0.1.0
djangorestframework==3.15.2
ez_setup==0.9
geographiclib==2.0
geopy==2.4.1
idna==3.10
jedi==0.17.2
numpy==1.24.4
panda==0.3.1
pandas==2.0.3
parso==0.7.1
pillow==10.3.0
pluggy==1.5.0
pyls==0.1.6
python-dateutil==2.9.0.post0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2025.2
rapidfuzz==3.9.7
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.9.0
urllib3==2.2.3